Gameloop GetBestTerrain
Rewrote get best terrain for standard loop to use a priority system. This will allow me to set precedence for terrain layers, making for smoother transitions.
The current approach relies on the level design working with the priority configuration. In situations where there are two terrain layers, Levels should highlight terrains with higher priorities as the "primary" layer, and terrains with lower priorities as the "secondary" layer
A more complex approach could give each terrain a priority based [0, 1], and accumulate totals for each terrain found in the vicinity. Finally it would return the terrain with the highest total. This would allow full control over blending, at maximum overhead cost.
The beauty of the former approach, is that it's realistic to most level designs I can think of. Any section with more than 2 terrain tiles should probably not rely heavily on the blending of those tiles anyways.


Former-commit-id: 3c6e4811571e1fd268f29501a7fc472fededf83f